+++ Engine/Render/Pipeline/BlitPipeline.cpp
@@ -4,7 +4,7 @@
 
 namespace ToolEngine
 {
-	BlitPipeline::BlitPipeline(RHIDevice& device, VkRenderPass render_pass): m_device(device), m_render_pass(render_pass)
+	BlitPipeline::BlitPipeline(RHIDevice& device, VkRenderPass render_pass): RenderPipeline(device, render_pass)
 	{
 		std::vector<RHIDescriptorType> layout_descriptor;
 		layout_descriptor.push_back(RHIDescriptorType::Sampler);
