+++ Engine/Render/Pipeline/RenderPipeline.cpp
@@ -1,0 +1,12 @@
+#include "RenderPipeline.h"
+
+namespace ToolEngine
+{
+	RenderPipeline::RenderPipeline(RHIDevice& device, VkRenderPass render_pass)
+		: m_device(device), m_render_pass(render_pass)
+	{
+	}
+	RenderPipeline::~RenderPipeline()
+	{
+	}
+}
